Extract excerpt caching code
